fix(database): fix broken ListTorrentByBangumi query, invalid Preload, and int/uint type inconsistency
- ListTorrentByBangumi queried non-existent columns on Torrent; rewrite
  to look up Bangumi first then filter by bangumi_id foreign key
- Remove Preload("BangumiParse") from GetTorrentWithDetails and
  ListTorrentWithDetails — Torrent has no such association field
- Unify primary key types: Bangumi.ID, EpisodeMetadata.ID/BangumiID,
  and Torrent.BangumiID changed from int to uint to match GORM convention
  and the uint already used in API layer; update all affected method
  signatures and parseIDs helper accordingly

+++ internal/database/bangumi_test.go
@@ -126,7 +126,7 @@ func TestBangumiLifecycle(t *testing.T) {
 
 	t.Run("GetWithDetails", func(t *testing.T) {
 		ctx := context.Background()
-		got, err := db.GetBangumiWithDetails(ctx, uint(bangumi.ID))
+		got, err := db.GetBangumiWithDetails(ctx, bangumi.ID)
 		if err != nil {
 			t.Fatalf("GetBangumiWithDetails failed: %v", err)
 		}
